requirements.txt
streamlit
pandas
numpy
python-dotenv
requests
matplotlib
